app/requirements.txt
flask==3.0.3
uwsgi==2.0.27

numpy==1.24.4
pillow==10.4.0

face-recognition==1.3.0
